+++ integration/cloud_account_test.go
@@ -17,7 +17,7 @@
 // limitations under the License.
 //
 
-package cloudAccount
+package integration
 
 import (
 	"testing"
@@ -34,19 +34,19 @@ func TestCloudAccountCommandAliases(t *testing.T) {
 
 	// lacework cloud-accounts
 	out, err, exitcode = LaceworkCLI("help", "cloud-accounts")
-	assert.Contains(t, out.String(), "lacework cloud-accounts [command]")
+	assert.Contains(t, out.String(), "lacework cloud-account [command]")
 	assert.Empty(t, err.String(), "STDERR should be empty")
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 
 	// lacework cloud
 	out, err, exitcode = LaceworkCLI("help", "cloud")
-	assert.Contains(t, out.String(), "lacework cloud [command]")
+	assert.Contains(t, out.String(), "lacework cloud-account [command]")
 	assert.Empty(t, err.String(), "STDERR should be empty")
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 
 	// lacework ca
 	out, err, exitcode = LaceworkCLI("help", "ca")
-	assert.Contains(t, out.String(), "lacework ca [command]")
+	assert.Contains(t, out.String(), "lacework cloud-account [command]")
 	assert.Empty(t, err.String(), "STDERR should be empty")
 	assert.Equal(t, 0, exitcode, "EXITCODE is not the expected one")
 }
@@ -103,7 +103,7 @@ func _TestCloudAccountCommandListWithTypeFlagErrorUnknownType(t *testing.T) {
 }
 
 func TestCloudAccountShow(t *testing.T) {
-	out, err, exitcode := LaceworkCLIWithTOMLConfig("cloud-account", "show", "TECHALLY_948AB76C4F809D5CBE4C92BB38F6EBFD9F413694FD85C75")
+	out, err, exitcode := LaceworkCLIWithTOMLConfig("cloud-account", "show", "TECHALLY_BA7DC3D27F930097DDD414B8FFF2F655FCB39075625F5E9")
 	// Summary Table
 	assert.Contains(t, out.String(), "CLOUD ACCOUNT GUID",
 		"STDOUT table headers changed, please check")